+++ packages/api-extractor/test/fixtures/export-default-exported-class.d.ts
@@ -1,0 +1,3 @@
+export class C {
+}
+export default C;