+++ packages/api-extractor/test/fixtures/export-default-exported-function.d.ts
@@ -1,0 +1,2 @@
+export function f(arg: any): void;
+export default f;