+++ packages/api-extractor/test/fixtures/extends-exported-base-class.d.ts
@@ -1,0 +1,12 @@
+export class C1 {
+    static prop1: boolean;
+    constructor(arg: boolean);
+    prop1: boolean;
+    method1(): void;
+}
+export class C2 extends C1 {
+    static prop2: boolean;
+    constructor(arg: number);
+    prop2: number;
+    method2(): void;
+}