+++ packages/api-extractor/test/fixtures/extends-hidden-base-class.d.ts
@@ -1,0 +1,9 @@
+export class C2 {
+    static prop2: boolean;
+    static prop1: boolean;
+    constructor(arg: number);
+    prop2: number;
+    method2(): void;
+    prop1: boolean;
+    method1(): void;
+}