+++ packages/api-extractor/test/fixtures/extends-hidden-generic-base-class.d.ts
@@ -1,0 +1,7 @@
+export class C2 {
+    constructor(arg: number);
+    prop2: number;
+    method2(): void;
+    prop1: boolean;
+    method1(arg: boolean): void;
+}